reverse_nxjson: remove insane allocation of 10000000000000 bytes
1e13 is 10000000000000, which doesn't even fit into an int.
instead, use a stack buffer of 128 kbyte which should be plenty,
then strdup the actual string on return.
note that this entire function is horribly inefficient as the
string has to be scanned over and over.
an efficient implementation would use

    len = snprintf(buf, sizeof buf "[%s]", s)

instead of

    strcat(s, "["), strcat(s, t), strcat(s, "]")

, and then add len to the start pointer.

while at it, also rename the confusing IF_... macro which suggests an
if block starts, and use a function-like macro.