Add default temperature thresholds
The tool to convert nbfc configs to json just left empty arrays for
temperature thresholds.

This PR fixes this by replacing thresholds with default values if
they are empty in the config. The defaults are taken from default
values in original nbfc.

I've also run the tool against the original nbfc tool to update the
configs.