Improved the forking and daemonization process
Previously, NBFC-Linux used threads to accept new connections and handle client requests.
Because of that, we needed to fork early and establish a communication channel
between the parent and child processes to confirm successful initialization.

Since NBFC-Linux now runs single-threaded, this synchronization is no longer
necessary, and the `pipe(), write(), read()` logic has been removed.

Additionally, the daemonization process was improved by calling `setsid()`
to create a new session and chdir("/") to change the working directory to the root.